Refactored routes
