+++ index.js
@@ -1,5 +1,5 @@
 const http = require("http");
-const { routes, notFoundRoute } = require("./routes");
+const routes = require("./routes/index");
 
 const server = http.createServer(async (req, res) => {
   const { url, method } = req;
@@ -8,7 +8,8 @@ const server = http.createServer(async (req, res) => {
       return route.sendResponse(req, res);
     }
   }
-  notFoundRoute.sendResponse(req, res);
+  res.writeHead(404, { "Content-Type": "text/plain" });
+  res.end("Invalid API Call");
 });
 
 server.listen(8000, () => {
