+++ routes/login.js
@@ -1,0 +1,14 @@
+const fs = require("fs");
+
+function loginGet(req, res) {
+  const data = fs.readFileSync("views/login.html");
+  const html = data.toString();
+  res.writeHead(200, {
+    "Content-Type": "text/html; charset=utf-8",
+    "X-Content-Type-Options": "nosniff",
+    "Content-Length": Buffer.byteLength(html, "utf-8"),
+  });
+  res.end(html);
+}
+
+module.exports = loginGet;
